Add filter command section in DG
Currently, the developer guide does not contain details about the filter command and relevant UML diagrams.

This needs to change because the DG is crucial to allow future developers to understand and build on existing code.

Let's update the existing developer guide to contain the overall details about the filter command.

Fixes #90